fix lingering restructuring errors
