+++ docs/source/pyskindose.helpers.rst
@@ -1,0 +1,21 @@
+pyskindose.helpers package
+==========================
+
+Submodules
+----------
+
+pyskindose.helpers.create\_attributes\_string module
+----------------------------------------------------
+
+.. automodule:: pyskindose.helpers.create_attributes_string
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: pyskindose.helpers
+   :members:
+   :undoc-members:
+   :show-inheritance:
