+++ docs/source/pyskindose.settings.rst
@@ -1,0 +1,85 @@
+pyskindose.settings package
+===========================
+
+Submodules
+----------
+
+pyskindose.settings.normalization\_settings module
+--------------------------------------------------
+
+.. automodule:: pyskindose.settings.normalization_settings
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyskindose.settings.patient\_offset module
+------------------------------------------
+
+.. automodule:: pyskindose.settings.patient_offset
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyskindose.settings.phantom\_dimensions module
+----------------------------------------------
+
+.. automodule:: pyskindose.settings.phantom_dimensions
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyskindose.settings.phantom\_settings module
+--------------------------------------------
+
+.. automodule:: pyskindose.settings.phantom_settings
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyskindose.settings.plot\_settings module
+-----------------------------------------
+
+.. automodule:: pyskindose.settings.plot_settings
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyskindose.settings.pyskindose\_settings module
+-----------------------------------------------
+
+.. automodule:: pyskindose.settings.pyskindose_settings
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyskindose.settings.rotation\_direction module
+----------------------------------------------
+
+.. automodule:: pyskindose.settings.rotation_direction
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyskindose.settings.translation\_direction module
+-------------------------------------------------
+
+.. automodule:: pyskindose.settings.translation_direction
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyskindose.settings.translation\_offset module
+----------------------------------------------
+
+.. automodule:: pyskindose.settings.translation_offset
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: pyskindose.settings
+   :members:
+   :undoc-members:
+   :show-inheritance:
